requirements.txt
fabric==1.2.2
coverage==3.5.1
django==1.3
numpy==1.6.1
pyquery==1.0
pil==1.1.7
django-activelink==0.3
django_compressor==1.1
beautifulsoup==3.2.0
cssmin==0.1.4
fixturefactory==0.2b
django-formwizard==1.0
